+++ baremaps-core/src/main/java/org/apache/baremaps/utils/MappedByteBufferUtils.java
@@ -10,7 +10,7 @@
  * the License.
  */
 
-package org.apache.baremaps.collection.utils;
+package org.apache.baremaps.utils;
