+++ baremaps-core/src/test/java/org/apache/baremaps/database/tile/FileTileStoreTest.java
@@ -18,7 +18,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import org.apache.baremaps.collection.utils.FileUtils;
+import org.apache.baremaps.utils.FileUtils;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
