+++ baremaps-core/src/test/java/org/apache/baremaps/iploc/IpLocTest.java
@@ -22,7 +22,6 @@ import java.nio.file.Paths;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
-import org.apache.baremaps.collection.utils.FileUtils;
 import org.apache.baremaps.geocoder.Geocoder;
 import org.apache.baremaps.geocoder.geonames.GeonamesGeocoder;
 import org.apache.baremaps.iploc.data.InetnumLocation;
@@ -35,6 +34,7 @@ import org.apache.baremaps.iploc.database.SqliteUtils;
 import org.apache.baremaps.iploc.nic.NicData;
 import org.apache.baremaps.iploc.nic.NicObject;
 import org.apache.baremaps.testing.TestFiles;
+import org.apache.baremaps.utils.FileUtils;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
